feat: 서브 이미지 클릭 시 이미지 변경 및 선택된 이미지 표시
